chapter_004/src/main/java/ru/job4j/iterator/EvenNumberIterator.java
package ru.job4j.iterator;

import java.util.Iterator;
import java.util.NoSuchElementException;

/**
 * @author Michael Hodkov
 * @version $Id$
 * @since 0.1
 */
public class EvenNumberIterator implements Iterator {
    private final int[] values;
    private int index = 0;

    public EvenNumberIterator(int[] values) {
        this.values = values;
    }

    @Override
    public boolean hasNext() {
        for (int i = index; i < values.length; i++) {
            if (values[i] % 2 == 0) {
                index = i;
                return true;
            }
        }
        return false;
    }

    @Override
    public Object next() {
        if (!hasNext()) {
            throw new NoSuchElementException("Not more even numbers");
        }
        return values[index++];
    }
}
